chore(): refactor types

+++ pkg/models/auth.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type Auth struct {
 	gorm.Model
-	Id       int32  `json:"id" gorm:"primaryKey"`
+	Id       int64  `json:"id" gorm:"primaryKey"`
 	Email    string `json:"email"`
 	Password string `json:"password"`
 }
